Refactored to include support for user-instantiated controllers.
