--- Sources/Vapor/Routing/ResourceController.swift
@@ -1,23 +1,0 @@
-/**
- * Organize your routing logic with a conformance of
- * `Controller`. Controls group related route logic into
- * a single protocol that, by default, conforms to standard
- * CRUD operations.
- */
-
-public protocol ResourceController: Controller {
-    /// Display many instances
-	func index(request: Request) throws -> ResponseConvertible
-
-    /// Create a new instance.
-    func store(request: Request) throws -> ResponseConvertible
-
-    /// Show an instance.
-    func show(request: Request) throws -> ResponseConvertible
-
-    /// Update an instance.
-    func update(request: Request) throws -> ResponseConvertible
-
-    /// Delete an instance.
-    func destroy(request: Request) throws -> ResponseConvertible
-}
